// tests/types/feelookup.types.ts
export type AmountType = 'Enviar' | 'Recibir';

export type MethodPayment = 'Efectivo' | 'Depósito' | 'Todos' | 'Wallet' | 'ATM' | 'PIX';

export type Currency =
  | 'USD'
  | 'COP'
  | 'BOB'
  | 'HTG'
  | 'ARP'
  | 'PEN'
  | 'BRL'
  | 'CAD'
  | 'EUR'
  | 'GBP'
  | 'CLP';

export type CurrencySelect = 'USD' | 'CLP';